Solved a massive leak in 'file_data_access.c' when reading actions. 'op_fields' variable was never free'd causing a total memory waste (each trace line stay in memory during the whole execution)
'MALLOC_*' (in 'mallocame.h|c') never used anymore.
All memory allocation now uses the pure standard
'malloc/free' calls.

All 'common-files' dependencies are now inside the
source code package. This is the first step in to rebuild
an actual 'common-files' in the Performance Tools team.

Almost all compilation 'warnings' cleaned.

+++ Simulator/model/task.h
@@ -119,9 +119,9 @@ t_boolean more_actions_to_sintetic (struct t_thread *thread);
 
 struct t_node *get_node_for_task_by_name (struct t_Ptask *Ptask, int taskid);
 
-void TASK_module_new (int    type,
-                             int    value,
-                             double ratio);
+void TASK_module_new (long long int type,
+                      long long int value,
+                      double        ratio);
 
 void file_name (struct t_Ptask *Ptask, int file_id, char *location);
